增加装备系统

+++ src/main/java/com/hz/dafeiji/ai/user/modules/equipments/Equipment.java
@@ -68,9 +68,9 @@ public class Equipment implements IdentityObj{
      * 从配置表构造新装备
      * @param templetId  模版ID
      */
-    public Equipment(int templetId){
+    public Equipment(long templetId){
         this.id = IdentityGen.INSTANCE.incrementAndGet();
-        this.templet = EquipmentTempletCfg.getEquipmentTempletById( templetId );
+        this.templet = EquipmentTempletCfg.getEquipmentTempletById( (int)templetId );
         this.level = 1;
         this.quality = templet.getQuality();
         this.loaded = 0;
